// frontend/src/api/generated/models/inbound-result-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipType } from './inbound-slip-type';
// May contain unused imports in some cases
// @ts-ignore
import type { UnitType } from './unit-type';

/**
 * 
 * @export
 * @interface InboundResultItem
 */
export interface InboundResultItem {
    /**
     * 入荷伝票ID
     * @type {number}
     * @memberof InboundResultItem
     */
    'slipId': number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof InboundResultItem
     */
    'slipNumber': string;
    /**
     * 
     * @type {InboundSlipType}
     * @memberof InboundResultItem
     */
    'slipType': InboundSlipType;
    /**
     * 仕入先コード
     * @type {string}
     * @memberof InboundResultItem
     */
    'partnerCode'?: string | null;
    /**
     * 仕入先名
     * @type {string}
     * @memberof InboundResultItem
     */
    'partnerName'?: string | null;
    /**
     * 明細ID
     * @type {number}
     * @memberof InboundResultItem
     */
    'lineId': number;
    /**
     * 明細行番号
     * @type {number}
     * @memberof InboundResultItem
     */
    'lineNo': number;
    /**
     * 商品コード
     * @type {string}
     * @memberof InboundResultItem
     */
    'productCode': string;
    /**
     * 商品名
     * @type {string}
     * @memberof InboundResultItem
     */
    'productName': string;
    /**
     * 
     * @type {UnitType}
     * @memberof InboundResultItem
     */
    'unitType': UnitType;
    /**
     * ロット番号
     * @type {string}
     * @memberof InboundResultItem
     */
    'lotNumber'?: string | null;
    /**
     * 賞味/使用期限日
     * @type {string}
     * @memberof InboundResultItem
     */
    'expiryDate'?: string | null;
    /**
     * 入荷予定数量
     * @type {number}
     * @memberof InboundResultItem
     */
    'plannedQty': number;
    /**
     * 検品数
     * @type {number}
     * @memberof InboundResultItem
     */
    'inspectedQty': number;
    /**
     * 差異数
     * @type {number}
     * @memberof InboundResultItem
     */
    'diffQty': number;
    /**
     * 入庫先ロケーションコード
     * @type {string}
     * @memberof InboundResultItem
     */
    'locationCode': string;
    /**
     * 入庫確定日時
     * @type {string}
     * @memberof InboundResultItem
     */
    'storedAt': string;
    /**
     * 入庫確定者名
     * @type {string}
     * @memberof InboundResultItem
     */
    'storedByName': string;
}