// frontend/src/api/generated/models/inbound-result-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InboundResultItem } from './inbound-result-item';

/**
 * 
 * @export
 * @interface InboundResultPageResponse
 */
export interface InboundResultPageResponse {
    /**
     * 
     * @type {Array<InboundResultItem>}
     * @memberof InboundResultPageResponse
     */
    'content': Array<InboundResultItem>;
    /**
     * 
     * @type {number}
     * @memberof InboundResultPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof InboundResultPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof InboundResultPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof InboundResultPageResponse
     */
    'totalPages': number;
}